backend/requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.12.0
incremental==21.3.0
Markdown==3.4.1
msgpack==1.0.4
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-dotenv==0.21.0
pytz==2022.2.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing-extensions==4.3.0
tzdata==2022.2
zipp==3.8.1
zope.interface==5.4.0
